Change _get_updates_for to yield parameter deltas.
Previously, _get_updates_for always yielded new parameter values. This
made the interface simpler in _get_updates, but it created unnecessary
computation (by having to remove the original parameter value again, in
cases where we were computing momentum). This change returns parameter
updates in the form of deltas, while updates for non-parameter
auxiliaries remain in the form of value replacements.

Added a couple small comments for guidance with this in _get_updates.